+++ projects/pastanaga-angular/src/lib/controls/textfield/select/select.component-form-control-name.spec.ts
@@ -8,68 +8,60 @@ import { initTest } from '../../form-field-test-utils.spec';
 import {
     testDescribedBy,
     testDisabledReactive,
-    testErrorMessage,
     testHelp,
     testId,
     testName,
     testNoErrorForPristine,
-    testPlaceholder,
     testReadonly,
 } from '../input/common-behaviors.spec';
 import { SelectComponent } from './select.component';
 import {
     TEST_OPTIONS,
-    testApplyUserInputInSuggestionMode,
     testErrorMessagesForSelect,
-    testFilterOptions,
-    testMaxlengthForSelect,
     testNoChangeForInactiveReactive,
     testParentAssignValueReactive,
-    testPatternForSelect,
     testRequiredForSelect,
     testSelectOption,
     testShowAllErrorsForSelect,
     testToggleDropDownWithInputClick,
-    testUserInputMatchOptionLabel,
-    testUserInputNotMatchingOptionLabel,
 } from './common-behaviors.spec';
 
-import {FormControl, FormGroup} from '@angular/forms';
+import { FormControl, FormGroup } from '@angular/forms';
 
 @Component({
     template: `
         <form [formGroup]="form">
-        <pa-select
-            #paInput
-            [id]="id"
-            [name]="name"
-            [help]="help"
-            [describedBy]="describedBy"
-            [readonly]="readonly"
-            [hasFocus]="hasFocus"
-            [showAllErrors]="showAllErrors"
-            [errorMessages]="errorMessages"
-            [errorMessage]="errorMessage"
-            [updateOn]="updateOn"
-            formControlName="text"
-            [suggestionMode]="suggestionMode"
-            [placeholder]="placeholder"
-            [required]="required"
-            [pattern]="pattern"
-            [label]="label"
-            [maxlength]="maxlength"
-            (valueChange)="onValueChange($event)"
-            (expanded)="onExpanded($event)"
-        >
-            <pa-option-header>Users</pa-option-header>
-            <pa-option value="user1">User 1</pa-option>
-            <pa-option value="user2">User 2</pa-option>
-            <pa-option value="user3">User 3</pa-option>
-            <pa-option-header>Groups</pa-option-header>
-            <pa-option value="group1" disabled>Group 1</pa-option>
-            <pa-option value="group2">Group 2</pa-option>
-            <pa-option value="group3">Group 3</pa-option>
-        </pa-select></form>
+            <pa-select
+                #paInput
+                [id]="id"
+                [name]="name"
+                [help]="help"
+                [describedBy]="describedBy"
+                [readonly]="readonly"
+                [hasFocus]="hasFocus"
+                [showAllErrors]="showAllErrors"
+                [errorMessages]="errorMessages"
+                [errorMessage]="errorMessage"
+                [updateOn]="updateOn"
+                formControlName="text"
+                [placeholder]="placeholder"
+                [required]="required"
+                [pattern]="pattern"
+                [label]="label"
+                [maxlength]="maxlength"
+                (valueChange)="onValueChange($event)"
+                (expanded)="onExpanded($event)"
+            >
+                <pa-option-header>Users</pa-option-header>
+                <pa-option value="user1">User 1</pa-option>
+                <pa-option value="user2">User 2</pa-option>
+                <pa-option value="user3">User 3</pa-option>
+                <pa-option-header>Groups</pa-option-header>
+                <pa-option value="group1" disabled>Group 1</pa-option>
+                <pa-option value="group2">Group 2</pa-option>
+                <pa-option value="group3">Group 3</pa-option>
+            </pa-select>
+        </form>
     `,
 })
 export class TemplateOptionsTestComponent {
@@ -90,18 +82,15 @@ export class TemplateOptionsTestComponent {
     pattern?: string | RegExp;
     maxlength?: number;
     label = 'the label';
-    suggestionMode = false;
     form = new FormGroup({
         text: new FormControl(),
     });
     get formControl() {
         return this.form.get('text') || new FormControl();
     }
-    onValueChange(event: any) {
-    }
+    onValueChange(event: any) {}
 
-    onExpanded(event: any) {
-    }
+    onExpanded(event: any) {}
 }
 
 describe('SelectComponent ngModel template options', () => {
@@ -109,7 +98,11 @@ describe('SelectComponent ngModel template options', () => {
     let fixture: ComponentFixture<TemplateOptionsTestComponent>;
     beforeEach(async(() => {
         nextId++;
-        fixture = initTest(TemplateOptionsTestComponent, SelectComponent, [PaIconModule, PaDropdownModule, PaPopupModule]);
+        fixture = initTest(TemplateOptionsTestComponent, SelectComponent, [
+            PaIconModule,
+            PaDropdownModule,
+            PaPopupModule,
+        ]);
     }));
 
     it('should assign id', fakeAsync(() => testId(fixture, nextId, 'select')));
@@ -129,15 +122,7 @@ describe('SelectComponent ngModel template options', () => {
 
     it('should not show errors when pristine', fakeAsync(() => testNoErrorForPristine(fixture)));
 
-    it('should apply errorMessage', fakeAsync(() => testErrorMessage(fixture)));
-
-    it('should assign placeholder', fakeAsync(() => testPlaceholder(fixture)));
-
     it('should validate required', fakeAsync(() => testRequiredForSelect(fixture)));
-
-    it('should validate pattern', fakeAsync(() => testPatternForSelect(fixture)));
-
-    it('should validate maxlength', fakeAsync(() => testMaxlengthForSelect(fixture)));
 
     it('should not propagate change when component is not active', fakeAsync(() =>
         testNoChangeForInactiveReactive(fixture, fixture.componentInstance.formControl)));
@@ -146,8 +131,6 @@ describe('SelectComponent ngModel template options', () => {
 
     it('should select option from template', fakeAsync(() => testSelectOption(fixture)));
 
-    it('should filter options when typing', fakeAsync(() => testFilterOptions(fixture)));
-
     it('should have parent value selected', fakeAsync(() => {
         testParentAssignValueReactive(fixture, fixture.componentInstance.formControl);
     }));
@@ -155,41 +138,34 @@ describe('SelectComponent ngModel template options', () => {
     it('should apply showAllErrors', fakeAsync(() => testShowAllErrorsForSelect(fixture)));
 
     it('should display errorMessages', fakeAsync(() => testErrorMessagesForSelect(fixture)));
-
-    it('should apply option value when user input match an option label', fakeAsync(() => testUserInputMatchOptionLabel(fixture)));
-
-    it('should keep actual value when user input does not match an option label',
-        fakeAsync(() => testUserInputNotMatchingOptionLabel(fixture)));
-
-    it('should apply user input on blur when suggestionMode is on', fakeAsync(() => testApplyUserInputInSuggestionMode(fixture)));
 });
 
 @Component({
     template: `
         <form [formGroup]="form">
-        <pa-select
-            #paInput
-            [options]="options"
-            [id]="id"
-            [name]="name"
-            [help]="help"
-            [describedBy]="describedBy"
-            [readonly]="readonly"
-            [hasFocus]="hasFocus"
-            [showAllErrors]="showAllErrors"
-            [errorMessages]="errorMessages"
-            [errorMessage]="errorMessage"
-            [updateOn]="updateOn"
-            formControlName="text"
-            [suggestionMode]="suggestionMode"
-            [placeholder]="placeholder"
-            [required]="required"
-            [pattern]="pattern"
-            [label]="label"
-            [maxlength]="maxlength"
-            (valueChange)="onValueChange($event)"
-            (expanded)="onExpanded($event)"
-        ></pa-select></form>
+            <pa-select
+                #paInput
+                [options]="options"
+                [id]="id"
+                [name]="name"
+                [help]="help"
+                [describedBy]="describedBy"
+                [readonly]="readonly"
+                [hasFocus]="hasFocus"
+                [showAllErrors]="showAllErrors"
+                [errorMessages]="errorMessages"
+                [errorMessage]="errorMessage"
+                [updateOn]="updateOn"
+                formControlName="text"
+                [placeholder]="placeholder"
+                [required]="required"
+                [pattern]="pattern"
+                [label]="label"
+                [maxlength]="maxlength"
+                (valueChange)="onValueChange($event)"
+                (expanded)="onExpanded($event)"
+            ></pa-select>
+        </form>
     `,
 })
 export class ParameterOptionsTestComponent {
@@ -210,7 +186,6 @@ export class ParameterOptionsTestComponent {
     pattern?: string | RegExp;
     maxlength?: number;
     label = 'the label';
-    suggestionMode = false;
     form = new FormGroup({
         text: new FormControl(),
     });
@@ -218,11 +193,9 @@ export class ParameterOptionsTestComponent {
         return this.form.get('text') || new FormControl();
     }
     options = TEST_OPTIONS;
-    onValueChange(event: any) {
-    }
+    onValueChange(event: any) {}
 
-    onExpanded(event: any) {
-    }
+    onExpanded(event: any) {}
 }
 
 describe('SelectComponent ngModel parameter options', () => {
@@ -230,7 +203,11 @@ describe('SelectComponent ngModel parameter options', () => {
     let fixture: ComponentFixture<ParameterOptionsTestComponent>;
     beforeEach(async(() => {
         nextId++;
-        fixture = initTest(ParameterOptionsTestComponent, SelectComponent, [PaIconModule, PaDropdownModule, PaPopupModule]);
+        fixture = initTest(ParameterOptionsTestComponent, SelectComponent, [
+            PaIconModule,
+            PaDropdownModule,
+            PaPopupModule,
+        ]);
     }));
 
     it('should assign readonly', fakeAsync(() => testReadonly(fixture)));
@@ -242,14 +219,6 @@ describe('SelectComponent ngModel parameter options', () => {
 
     it('should not show errors when pristine', fakeAsync(() => testNoErrorForPristine(fixture)));
 
-    it('should apply errorMessage', fakeAsync(() => testErrorMessage(fixture)));
-
-    it('should validate required', fakeAsync(() => testRequiredForSelect(fixture)));
-
-    it('should validate pattern', fakeAsync(() => testPatternForSelect(fixture)));
-
-    it('should validate maxlength', fakeAsync(() => testMaxlengthForSelect(fixture)));
-
     it('should not propagate change when component is not active', fakeAsync(() =>
         testNoChangeForInactiveReactive(fixture, fixture.componentInstance.formControl)));
 
@@ -257,7 +226,8 @@ describe('SelectComponent ngModel parameter options', () => {
 
     it('should select option from template', fakeAsync(() => testSelectOption(fixture)));
 
-    it('should have parent value selected', fakeAsync(() => testParentAssignValueReactive(fixture, fixture.componentInstance.formControl)));
+    it('should have parent value selected', fakeAsync(() =>
+        testParentAssignValueReactive(fixture, fixture.componentInstance.formControl)));
 
     it('should apply showAllErrors', fakeAsync(() => testShowAllErrorsForSelect(fixture)));