+++ projects/pastanaga-angular/src/lib/controls/textfield/select/select.component-form-control.spec.ts
@@ -8,33 +8,25 @@ import { initTest } from '../../form-field-test-utils.spec';
 import {
     testDescribedBy,
     testDisabledReactive,
-    testErrorMessage,
     testHelp,
     testId,
     testName,
     testNoErrorForPristine,
-    testPlaceholder,
     testReadonly,
 } from '../input/common-behaviors.spec';
 import { SelectComponent } from './select.component';
 import {
     TEST_OPTIONS,
-    testApplyUserInputInSuggestionMode,
     testErrorMessagesForSelect,
-    testFilterOptions,
-    testMaxlengthForSelect,
     testNoChangeForInactiveReactive,
     testParentAssignValueReactive,
-    testPatternForSelect,
     testRequiredForSelect,
     testSelectOption,
     testShowAllErrorsForSelect,
     testToggleDropDownWithInputClick,
-    testUserInputMatchOptionLabel,
-    testUserInputNotMatchingOptionLabel,
 } from './common-behaviors.spec';
 
-import {FormControl, FormGroup} from '@angular/forms';
+import { FormControl } from '@angular/forms';
 
 @Component({
     template: `
@@ -51,7 +43,6 @@ import {FormControl, FormGroup} from '@angular/forms';
             [errorMessage]="errorMessage"
             [updateOn]="updateOn"
             [formControl]="formControl"
-            [suggestionMode]="suggestionMode"
             [placeholder]="placeholder"
             [required]="required"
             [pattern]="pattern"
@@ -89,14 +80,11 @@ export class TemplateOptionsTestComponent {
     pattern?: string | RegExp;
     maxlength?: number;
     label = 'the label';
-    suggestionMode = false;
     formControl = new FormControl();
 
-    onValueChange(event: any) {
-    }
+    onValueChange(event: any) {}
 
-    onExpanded(event: any) {
-    }
+    onExpanded(event: any) {}
 }
 
 describe('SelectComponent ngModel template options', () => {
@@ -104,7 +92,11 @@ describe('SelectComponent ngModel template options', () => {
     let fixture: ComponentFixture<TemplateOptionsTestComponent>;
     beforeEach(async(() => {
         nextId++;
-        fixture = initTest(TemplateOptionsTestComponent, SelectComponent, [PaIconModule, PaDropdownModule, PaPopupModule]);
+        fixture = initTest(TemplateOptionsTestComponent, SelectComponent, [
+            PaIconModule,
+            PaDropdownModule,
+            PaPopupModule,
+        ]);
     }));
 
     it('should assign id', fakeAsync(() => testId(fixture, nextId, 'select')));
@@ -117,19 +109,12 @@ describe('SelectComponent ngModel template options', () => {
 
     it('should assign readonly', fakeAsync(() => testReadonly(fixture)));
 
-    it('should apply disabled state', fakeAsync(() => testDisabledReactive(fixture, fixture.componentInstance.formControl)));
+    it('should apply disabled state', fakeAsync(() =>
+        testDisabledReactive(fixture, fixture.componentInstance.formControl)));
 
     it('should not show errors when pristine', fakeAsync(() => testNoErrorForPristine(fixture)));
 
-    it('should apply errorMessage', fakeAsync(() => testErrorMessage(fixture)));
-
-    it('should assign placeholder', fakeAsync(() => testPlaceholder(fixture)));
-
     it('should validate required', fakeAsync(() => testRequiredForSelect(fixture)));
-
-    it('should validate pattern', fakeAsync(() => testPatternForSelect(fixture)));
-
-    it('should validate maxlength', fakeAsync(() => testMaxlengthForSelect(fixture)));
 
     it('should not propagate change when component is not active', fakeAsync(() =>
         testNoChangeForInactiveReactive(fixture, fixture.componentInstance.formControl)));
@@ -138,9 +123,8 @@ describe('SelectComponent ngModel template options', () => {
 
     it('should select option from template', fakeAsync(() => testSelectOption(fixture)));
 
-    it('should filter options when typing', fakeAsync(() => testFilterOptions(fixture)));
-
-    it('should have parent value selected', fakeAsync(() => testParentAssignValueReactive(fixture, fixture.componentInstance.formControl)));
+    it('should have parent value selected', fakeAsync(() =>
+        testParentAssignValueReactive(fixture, fixture.componentInstance.formControl)));
 
     it('should apply showAllErrors', fakeAsync(() => testShowAllErrorsForSelect(fixture)));
 
@@ -163,7 +147,6 @@ describe('SelectComponent ngModel template options', () => {
             [errorMessage]="errorMessage"
             [updateOn]="updateOn"
             [formControl]="formControl"
-            [suggestionMode]="suggestionMode"
             [placeholder]="placeholder"
             [required]="required"
             [pattern]="pattern"
@@ -192,14 +175,11 @@ export class ParameterOptionsTestComponent {
     pattern?: string | RegExp;
     maxlength?: number;
     label = 'the label';
-    suggestionMode = false;
     formControl = new FormControl();
     options = TEST_OPTIONS;
-    onValueChange(event: any) {
-    }
+    onValueChange(event: any) {}
 
-    onExpanded(event: any) {
-    }
+    onExpanded(event: any) {}
 }
 
 describe('SelectComponent ngModel parameter options', () => {
@@ -207,22 +187,19 @@ describe('SelectComponent ngModel parameter options', () => {
     let fixture: ComponentFixture<ParameterOptionsTestComponent>;
     beforeEach(async(() => {
         nextId++;
-        fixture = initTest(ParameterOptionsTestComponent, SelectComponent, [PaIconModule, PaDropdownModule, PaPopupModule]);
+        fixture = initTest(ParameterOptionsTestComponent, SelectComponent, [
+            PaIconModule,
+            PaDropdownModule,
+            PaPopupModule,
+        ]);
     }));
 
     it('should assign readonly', fakeAsync(() => testReadonly(fixture)));
 
-    it('should apply disabled state', fakeAsync(() => testDisabledReactive(fixture, fixture.componentInstance.formControl)));
+    it('should apply disabled state', fakeAsync(() =>
+        testDisabledReactive(fixture, fixture.componentInstance.formControl)));
 
     it('should not show errors when pristine', fakeAsync(() => testNoErrorForPristine(fixture)));
-
-    it('should apply errorMessage', fakeAsync(() => testErrorMessage(fixture)));
-
-    it('should validate required', fakeAsync(() => testRequiredForSelect(fixture)));
-
-    it('should validate pattern', fakeAsync(() => testPatternForSelect(fixture)));
-
-    it('should validate maxlength', fakeAsync(() => testMaxlengthForSelect(fixture)));
 
     it('should not propagate change when component is not active', fakeAsync(() =>
         testNoChangeForInactiveReactive(fixture, fixture.componentInstance.formControl)));
@@ -237,11 +214,4 @@ describe('SelectComponent ngModel parameter options', () => {
     it('should apply showAllErrors', fakeAsync(() => testShowAllErrorsForSelect(fixture)));
 
     it('should display errorMessages', fakeAsync(() => testErrorMessagesForSelect(fixture)));
-
-    it('should apply option value when user input match an option label', fakeAsync(() => testUserInputMatchOptionLabel(fixture)));
-
-    it('should keep actual value when user input does not match an option label',
-        fakeAsync(() => testUserInputNotMatchingOptionLabel(fixture)));
-
-    it('should apply user input on blur when suggestionMode is on', fakeAsync(() => testApplyUserInputInSuggestionMode(fixture)));
 });